Create project_1.py
The first project is implementing a common strategy and  is generating a trading portfolio. Revision are make to include artificial intelligence and machine learning to boost the existing strategy. Key performance indicator is included to test the hypothesis and draw the conclusion to deploy the strategy to the real market.