Create project_2.py
The second project is a modification of the first project's strategy. It is still a common strategy within finance community. I am revised to include the artificial intelligence. We will test the first hypothesis and the second hypothesis. The first hypothesis is if our current strategy is no better than a random buy and sell actions. The second hypothesis is even if we add the artificial intelligence, our modified strategy is no better than a random walk. We are to finding out to accept the null hypothesis or to reject our null hypothesis.